Fix the <Return> binding for tk_dialog to invoke the focused button.
The standard for Windows and Gtk+ is that the <Return> binding invokes
the button that currently has the keyboard focus. This can be seen in
tk_messageBox which implements this.
This patch implements this feature and also makes use of the tk grab
utility functions to replace the  grab handling code.


FossilOrigin-Name: 8f0a215fbfd4a977ec7a1669bffd68a047318e89