TIP 360: Remove special handling of .help menu on X11.
Tk is the only toolkit to continue the Motif scheme of anchoring a Help
menu to the extreme right of the menubar. This patch disables the special
handling of the .help menu so that the help menu will be grouped with the
other menu items as is common with other toolkits.

The old behaviour can be restored by setting the option *Menu.useMotifHelp
to true or by calling 'tk::classic::restore menu'.

Signed-off-by: Pat Thoyts <patthoyts@users.sourceforge.net>


FossilOrigin-Name: 6821c28743dad60517ae8e5ec511dd94962a8659