TIP #359: Extended window manager hints support for X11.
Modern unix window managers use a set of window properties to give
hints as to the purpose of a toplevel window. They then use these
hints to apply various animation and decoration options based on the
type (dialog, menu, tooltip and more).

This patch adds a [wm attributes $w -type] option to control and read
the type hint and makes use of this for the ttk::combobox and the
dialogs raised from the Tk library scripts.

Signed-off-by: Pat Thoyts <patthoyts@users.sourceforge.net>


FossilOrigin-Name: 93a297dd4b38bfb111f2f4208a274213de88aacd