[Bug 3205260] avoid crash in wm manage of children with null class
The tkhtml widget doesn't set it's window class. The Windows implementation
of the wm manage command checks to see if the contained window is actually
a menu (looking for torn-off menus) by examining the class. This patch
adds a check for a null class to avoid crashing as suggested.

Suggested-by: Eric Boudaillier <beric@users.sourceforge.net>
Signed-off-by: Pat Thoyts <patthoyts@users.sourceforge.net>

FossilOrigin-Name: 27afdfb6d53bbf9ba79fd89024a36405cffdf824