[Bug 3205464] - handle [wm forget] when the parent is unmapped on windows
The fix for [Bug 2009788] prevented a crash but causes windows to be lost
if their parent is unmapped when [wm forget] is called. Added a test for
this case and ensure that the parent window exists when we remap the
child to its parent.

Reported-by: Koen Danckaert <danckaert@users.sourceforge.net>
Reported-by: Eric Boudaillier <beric@users.sourceforge.net>
Signed-off-by: Pat Thoyts <patthoyts@users.sourceforge.net>

FossilOrigin-Name: ecfa48e76b12e6868e56a31c7fc26e8aa2e3b194