[Bug # 3306909]: tk_popup placement on Windows Vista can be incorrect.
When posting menus near the bottom of the screen on Windows versions
newer than XP the menu may be placed incorrectly. This is due to code
in PostOverPoint that corrects an error in the XP window manager. This
fix is no longer required for Vista and Windows7.

Reported-by: Harald Oehlmann <oehhar@users.sourceforge.net>
Signed-off-by: Pat Thoyts <patthoyts@users.sourceforge.net>

FossilOrigin-Name: 6db4a05c4ecb22120951316412887f21b9d389b9