[3126428] Repaint ttk labels and buttons when the image is changed.
In Tk the images associated with labels and buttons have their image
changed callback set to cause the widget to be redrawn if the image is
changed in any way. However, this has not been done for the ttk equivalent
widgets.

FossilOrigin-Name: 70a23e858e3e7e79f4c87dc6018495bdab666aaa