Aqua X11 emulation: support Retina window source
Existing programs/libraries expect XGetImage() to return either an
XImage with the requested width and height, or NULL. This means that
a capture of a real window on a Retina display must be downscaled
before being returned by XGetImage() for compatibility.

To support this, a parameter force_1x_scale is added to
CreateCGImageFromDrawableRect(), which if true will downscale the
returned CGImage to the requested width and height. Otherwise,
capturing a window on a Retina display will return a CGImage having
twice the requested width and height.

If programs should be able to retrieve unscaled captures of windows on
Retina displays, then a separate API is needed. See
https://core.tcl-lang.org/tk/info/fcd6717d8b

Not fully tested. Would be particularly interested in a way to test
XCopyArea() between non-pixmaps. XCopyPlane() might be unfinished.