Aqua: use clipping to draw ttk::spinbox up/down buttons separately
I noticed the issue reported in [c36e462d]. I agree with Csaba’s
suggested change to avoid the moving arrows, and that is likely good
enough. Under a screen magnifier, however, there is still a subtle but
noticeable effect where the button appears slightly taller/larger when
the down arrow is pressed. I think this is due to how pressing the
down arrow requires the entire button to be drawn twice.

This effect can be avoided by using clipping to draw the up and down
arrow buttons separately, so that each is only drawn once more like
how Ttk expects. There is already some clipping code in place to
achieve this, though only on newer macOS where HIToolbox is not used.